src/transfermarkt_competicoes/requirements.txt
bs4==4.9.0
google-auth==2.12.0
google-auth-oauthli==0.5.3
pyyaml==6.0
gscpread==5.6.0
pandas==1.5.0
requests==2.28.1